imagersite/imager_images/views.py
#!/usr/bin/env python
# -*- coding: utf-8 -*-

from __future__ import unicode_literals
# from django.views.generic.base import TemplateView
from django.shortcuts import get_object_or_404
from django.views.generic import ListView, DetailView
from imager_images.models import Photo, Album
from braces.views import LoginRequiredMixin


class LibraryView(LoginRequiredMixin, ListView):
    context_object_name = 'media'
    template_name = 'imager_images/library.html'

    def get_queryset(self):
        albums = self.request.user.albums.all()
        photos = self.request.user.photos.all()
        return [albums, photos]


class AlbumView(LoginRequiredMixin, ListView):
    context_object_name = 'media'
    template_name = 'imager_images/album.html'

    def get_queryset(self):
        album = get_object_or_404(Album, id=self.kwargs['pk'])
        photos = album.photos.all()
        return [album, photos]


class PhotoView(LoginRequiredMixin, DetailView):
    model = Photo
    context_object_name = 'photo'
    template_name = 'imager_images/photo.html'
