requirements.txt
Django==1.8.3
django-bootstrap3==6.1.0
django-braces==1.8.1
django-registration-redux==1.2
Factory==1.2
factory-boy==2.5.2
fake-factory==0.5.2
gnureadline==6.3.3
ipython==3.2.1
Pillow==2.9.0
selenium==2.46.1
six==1.9.0
sorl-thumbnail==12.3
splinter==0.7.2
wheel==0.24.0
